ch11-app-decompose/ch11/requirements.txt
anyio==3.6.1
asgiref==3.5.2
asyncio==3.4.3
circuitbreaker==1.3.2
click==8.1.3
colorama==0.4.4
Deprecated==1.2.13
Django==3.2
dnspython==2.2.1
fastapi==0.78.0
Flask==2.1.2
googleapis-common-protos==1.56.3
grpcio==1.47.0
h11==0.13.0
idna==3.3
ifaddr==0.2.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
motor==3.0.0
opentelemetry-api==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-exporter-jaeger-proto-grpc==1.11.1
opentelemetry-exporter-jaeger-thrift==1.11.1
opentelemetry-instrumentation==0.31b0
opentelemetry-instrumentation-asgi==0.31b0
opentelemetry-instrumentation-fastapi==0.31b0
opentelemetry-instrumentation-logging==0.31b0
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
opentelemetry-util-http==0.31b0
opentracing==2.4.0
prometheus-client==0.14.1
protobuf==4.21.1
py-eureka-client==0.11.3
pydantic==1.9.1
pymongo==4.1.1
pytz==2022.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.1
starlette-exporter==0.13.0
thrift==0.16.0
typing_extensions==4.2.0
uvicorn==0.17.6
Werkzeug==2.1.2
win32-setctime==1.1.0
wrapt==1.14.1
zipp==3.8.0
